ToolsUI de-clutter.  Start splitting off OpPanels.
