+++ cdm/src/main/java/ucar/nc2/dt/radial/Nexrad2RadialAdapter.java
@@ -11,8 +11,8 @@ import ucar.nc2.constants.FeatureType;
 import ucar.nc2.dt.*;
 import ucar.nc2.ft.FeatureDataset;
 import ucar.nc2.time.CalendarDateUnit;
-import ucar.nc2.units.DateUnit;
 import ucar.nc2.units.DateFormatter;
+import ucar.nc2.units.DateUnit;
 import ucar.ma2.*;
 import ucar.nc2.Attribute;
 import ucar.unidata.geoloc.LatLonRect;
