+++ ui/src/main/java/ucar/nc2/ui/opp/CoordSysTable.java
@@ -1,9 +1,9 @@
 /*
- * Copyright (c) 1998-2018 University Corporation for Atmospheric Research/Unidata
+ * Copyright (c) 1998-2019 University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
 
-package ucar.nc2.ui;
+package ucar.nc2.ui.opp;
 
 import ucar.ma2.*;
 import ucar.ma2.DataType;
